Moved business logic to `schedule.ts`
